Migrate Glide to default to ALWAYS_ARGB_8888
Due to issues where we were seeing Camera builds request 8888 and
not always get it, employ a simple fix to just change Glide defaults
for Android across the board to 8888. The BitmapDecoders are what
really matter for Camera, changing the Builder as well for consistency.

Also added cache debugging lines to Engine to characterize issue.

Bug: 17523927
Change-Id: I10827561d0dbfc9e22365cceee16e030329071ff